refactor(hooks): :boom: remove index and use direct imports
