+++ src/__tests__/useFileSimulation.test.ts
@@ -1,6 +1,6 @@
 /** @jest-environment jsdom */
 import { renderHook, act } from '@testing-library/react';
-import { useFileSimulation } from '../client/hooks';
+import { useFileSimulation } from '../client/hooks/useFileSimulation';
 import { createFileSimulation } from '../client/fileSimulation';
 
 jest.mock('../client/fileSimulation');
